pyproject.toml
[tool.poetry]
name = "statusburo"
version = "0.1.0"
description = ""
authors = ["sloev <johanned.valbjorn@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
sanic = "^20.9.0"
pyjwt = "^1.7.1"
aiosqlite = "^0.15.0"
black = "^20.8b1"
aiohttp = "^3.6.2"
trustpilot_json_logging = "^1.0.12"
trustpilot-json-logging = "^1.0.12"
pillow = "^7.2.0"
timeago = "^1.0.14"
python-dateutil = "^2.8.1"
pygifsicle = "^1.0.1"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
start = 'statusburo.app:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

